Fix cassettes and helper so all tests work together
The search tests use a fake API profile which is *not* being used to
filter sensitive data - it's just fake. The record tests want a real
API profile, but it can't be 'apiwhatnot' plus filter_sensitive_data,
because doing so breaks the search tests. So now the record tests have
their own fake, filtered profile, which does not interact with the
search test cassettes.