Adds expiring URL support
EDS API sometimes returns expiring URLs. To allow users to have a
pleasant experience, we suppress display of those URLs until the users
clicks the "View Online" button at which point we ask for a _new_ copy
of the expiring URL and then redirect the user to that non-expired
version.

Also includes support to ensure guests are not taken to expiring URLs
as they are not authenticated on the target system.