Suppress display of request scan button for engineering standards
It's irritating to do this by subject header, since that isn't in the
Aleph data. We could pass in the EBSCO subject data, clean it (it's all
in searchLink format), and hope that some regex on that matches our
needs...

...but we also have a spreadsheet of all the standards we don't scan,
which includes their Barton item numbers, and that's data we DO have
from Aleph. Woot!