Exclude scanning button based on Subject Headings
The initial work on this feature used a list of Aleph IDs as the
exclusion criteria. After business rules were clarified, it became
desirable to use the same logic we use elsewhere in our systems,
which is the LC Subject Headings of the record.

This implements the business logic defined in:
https://mitlibraries.atlassian.net/browse/DI-592

This completes the work in:
https://mitlibraries.atlassian.net/browse/DI-608

Note: We pass in `oclc` and `scan` to item looksups not because we love
it, but because it saves us from making two API calls to Aleph instead
of one. This is a case of efficiency winning out over a preferred
application design