Provides fallback link to SFX menu if we don't know what else to do
* Refactors availability section.

There's already a lot of logic going on in the view to decide what to
show the user under Availability, and DI-607 requires additional logic.
This is too heavy to belong in a view and now lives in the helper.

* Refactors SFX URL generation out of ButtonMaker

We'll need to also access SFX URL generation in record_helper.rb, so it
needs to live in a shared place that understands SFX URL construction
more generally.

* Adds generic SFX link when all other, better-specified, options fail.

* Adds test coverage