Check fulltext link against ezproxy
If we have a fulltext link that is not known to be SFX or Ebsco (where
authentication is handled on the other end), check to see if we can
proxy it before sending the user fruitlessly to the proxy server.

This is only enabled for non-Aleph records because some of our Aleph
records have proxyable links that don't lead to full text (for example
to Table of Contents pages).

Tests are not complete as this is WIP awaiting stakeholder feedback on
whether it would be useful given our limitations.

https://mitlibraries.atlassian.net/browse/DI-633