Updating coveralls gem
Oddly, coveralls and the underlying simplecov library were in conflict
in versions when they were both added to the Gemfile.

This removes the explicit simplecov and uses the simplecov version that
coveralls requires which allows coveralls to use a newer version which
doesn’t include a different out of date library. <sigh>