Handle different failed EDS record responses
We were handling all EDS record failures the same, but we have seen two distinct cases in production so this handles those more clearly. We will also continue to handle the generic case to ensure we are seeing in our logs when all of these types happen and respond as appropriate.

https://mitlibraries.atlassian.net/browse/DI-721