Enables rack timeout
Default is 15 seconds. Can be controlled via ENV settings.
See https://github.com/heroku/rack-timeout for details.

Background: we are seeing a lot of Heroku H12 timeouts which are not really good. This should avoid that while not specifically addressing _why_ we see so many request timeouts. Basically this will allow us to timeout the request in such a way that the request terminates rather than heroku just stopping paying attention to it.