cache auth token and session token
We currently perform 8 API calls to EDS for every search on our site.

This reduces that number to 2 by caching the Auth Token (EDS Docs suggest caching this for 30 minutes) and session tokens (EDS Docs suggest caching this per user, but for our bento view we have no concept of users so this should be fine... for record view we use their gem and this logic is not in place).

With this we also don't call end session any more as we'll continue to use the session until close to it's expiration time on their end so we'll let them do their own clean up.

TLDR: Once every 25 minutes, a search will require up to 8 API calls like they all do now. Between those times, most searches will require 2 (one for books, one for articles).

The expectation is that by making fewer calls to the EDS API we will have fewer opportunities for timeouts in making calls to the EDS API and a net result of a better user experience while also putting less load on both our server and the EDS servers.

We also remove retry attempts (I'm not sure these were working at all based on logz reporting... if we need them after this work is complete, it'll be better to rethink them separately).