Adds dotenv-rails
We have relied on the `heroku local web` command to run the app locally and load from `.env`. This allows the option of using the standard `rails server` command. `heroku local` still works fine.

The main reason for this is that `rails server` allows for easily using `binding.pry` (or `byebug` but this project uses pry) whereas the `heroku local` makes that more difficult.