Adds Aeon request buttons
This adds two button options for Aleph results that, if the metadata for
the record allows, will provide a two links through SFX to Aeon.

https://mitlibraries.atlassian.net/browse/AIP-362

Business logic:
https://mitlibraries.atlassian.net/wiki/spaces/AIP/pages/989560879/Aeon+Request+Buttons+in+Barton

This also removes the no longer needed Contact Us button we previously
displayed for some of the records that now get Aeon links.