Updates eds-ruby
Updates eds-ruby and adjusts tests and cassettes to match the new way
the gem works. Unfortunately, some new features we don't use seem to
make many more additional calls per each record in our full record view
by default.

For now we can roll with this but it's super inefficient and unfortunate
this is how their official wrapper gem works.

A couple tests had manual cassette updates and are noted in the code as
things to be wary of in future maintenance.