Minor syntax changes
this commit updates cassettes to match a new base URL.

A follow on commit will regenerate them to ensure we are testing against up to date API responses. I want to make multiple commits to allow us to more closely monitor the changes taking place here for concerns.

FWIW. As of this point, the app appears to be working.