All items are hold status enabled if recalls are off
I'm not entirely sure yet if this is the way to go with this but we need stakeholder input on it.

The idea is that in the before times, we had either a Hold or Recall button depending on the status of the item. Never both. If checked out / not available... recall. However, now we want to allow Holds even on unavailable items... one day this will want to be reversible and I think we can just do this simple config to couple the logic to one setting. Maybe.

https://mitlibraries.atlassian.net/browse/DI-917