All request digital for available reserve items
Status 24 items (high use, usually on resevre) not eligible for
Request Item, but need to be eligible for Request Digital Copy

https://mitlibraries.atlassian.net/browse/DI-932

Also removes the note to ask at desk for reserve items.

Adds an additional collection name for reserves but leaves the old one
as it seems like it'll return one day.