Update pull request template
Why these changes are being introduced:

We've changed the way we structure and commit messages in recent
projects. Now that we use commit message templates, the current Bento
PR template asks for much of the information that is now included in
our commit message templates.

Relevant ticket(s):

None.

How this addresses that need:

This updates the PR template to match what we currently use in Thing.

Side effects of this change:

None.