Add Primo Search model
Why these changes are being introduced:

With the Primo launch forthcoming, we need to update Bento to query the
Primo Search API.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/IMP-1871

How this addresses that need:

This introduces a Primo Search model, which includes some rudimentary
logic to call the Primo Search API, and updates the readme accordingly.

Side effects of this change:

* The db schema timestamp updated in a weird way, as a result of my
having to run migrations locally.
* There are two new VCR cassettes: one for a normal Primo search, and
one for a bad Primo API response.
* Various API params have been added to ENV. Notably, PRIMO_VID_INST
is actually two params (see the changes to the readme for more info).