Enable scoped Primo searches
Why these changes are being introduced:

Our current implementation of the Search Primo model does not allow us
to scope searches to Alma or CDI.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/IMP-1961

How this addresses that need:

This adds a `scope` parameter to the query string constructor in the
Search Primo model, into which we can pass the `alma` (Books+) or
`cdi` (Articles+) scopes. These scopes are implemented as ENV variables,
so we can change them as needed. It also changes the tab to the new
`bento` tab, which isn't strictly necessary but could save us trouble
in the future.

Side effects of this change:

* The `inst` query param, which is not required for hosted customers, has
been removed, and the variable name changed from `PRIMO_VID_INST` to
`PRIMO_VID`.
* Removed the `popcorn_primo` cassette and added scoped cassettes
(`popcorn_primo_articles` and `popcorn_primo_books`).