Add normalization models for Primo searches
Why these changes are being introduced:

The app can currently search Primo and return raw results, but it cannot
normalize those results and map them into our Result model.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/IMP-1953

How this addresses that need:

This adds the relevant normalization models for Primo to map Primo search
results (both Alma and CDI) into our Result model to render in the result
partial. Primo searches can now be seen in the search_boxed view using
the 'alma' and 'cdi' targets.

Side effects of this change:

* Ex Libris has minimal useful documentation on the Primo Search API, so
this code will likely need to be refined over time.
* In an effort to keep things as simple as possible, this changeset
largely mirrors the logic for EDS normalization and minimizes changes
to the Result model and corresponding views. Because of this, some of the
code is a little more convoluted than if we were starting from scratch.
* We are constructing thumbnail URLs to Syndetics using the ISBN. A
PNX record can have multiple ISBNs, so we are taking the first entry
for now.
* All PNX fields are returned as arrays, so it is unclear which are
supposed to be multi-value and which are single-value. We are making some
assumptions to this effect.