Add 'Get It' links for Primo results.
Why these changes are being introduced:

If a resource is available online, a user should be able to access
the full text from the Bento record.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/IMP-2008

How this addresses that need:

This renders full-text/Get It links when available for Alma and CDI
results. For Alma results, we construct an Alma link resolver URL if the
resource is tagged as 'Alma-E'. CDI results already have these URLs in
the API response, so we can pull them directly from the JSON.

Side effects of this change:

Because some Alma records are not yet linked to their eresources, certain
'Get It' links will not be proxied. This is a known issue and is being
actively investigated.