Fix author linking bug
Why these changes are being introduced:

The Primo API occasionally provides multiple author names in a single
array element. The Primo normalization code reads this as a single author
name and generates a single link to the Primo author browse results,
instead of one link per author. This bug was identified during UX revivew
of IMP-1953.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/IMP-1953

How this addresses that need:

This adds some logic to our author sanitization code to look for more
bad data and attempt to fix it.

Side effects of this change:

None, but there are likely more examples of bad Ex Libris data requiring
remediation that we've yet to find. We will continue to push bug fixes
and add regression testing as needed.