Implement stakeholder feedback for IMP-2008
Why these changes are being introduced:

* Full-text links for CDI records lead to a disambiguation page with no
styling.
* Some records' links include a link to the eresource's table of contents,
but not to the full-text.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/IMP-2008

How this addresses that need:

* Modifies the openurl call in the `almaOpenUrl` field to redirect to the
Primo UI.
* Adds `u.ignore_date_coverage` param to all openurl links so full-text
links display instead of ToC links.

Side effects of this change:

This assumes consistency in the `almaOpenUrl` field in the Primo Search API response. If a CDI record does not have the standard base URL in this
field, then the full-text link will likely fail.