Remove hyphens from subject links
Why these changes are being introduced:

Relevance in the Primo subject browse is poor for queries that include
hyphens. This is a problem because compound subjects in library metadata
are divided with hyphens.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/IMP-2056

How this addresses that need:

This removes hyphens from compound subject links.

Side effects of this change:

We are still displaying hyphens in the UI, so there's a risk for
confusion when the linked Primo subject browse query is punctuated
differently.