Change author links to regular searches
Why these changes are being introduced:

Primo browse search does not include CDI results, which will be
confusing for Bento author links, especially when clicking an author link
on a CDI record.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/IMP-2072

How this addresses that need:

This reverts back to using regular Primo searches instead of browse
searches for author links.

Side effects of this change:

We are using the `exact` operator instead of `contains`, so the linked
search will include fewer results (but will be more accurate).