Display record-level real-time availability
Why these changes are being introduced:

Our original plan was to use the Alma Bib API for RTA, but we are
considering whether a more lightweight option using the Primo Search API
is viable. This scaled back version of RTA will tell users whether any
physical copy is available for a given result.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/IMP-2093

How this addresses that need:

* Updates the normalize_primo_books model to pull availability status
from catalog records and pass it into the result model.
* Adds availability? method to the result model that determines whether
any copies of a result are available for checkout.
* Adds check_holdings? method to the result model, which corresponds
with partial/unclear availability (e.g., one or more volumes is checked
out).
* Updates result view to display availability status.

Side effects of this change:

Minor CSS changes to clean up display of the new data.