Add feature flag to redirect full record links to Primo
Why these changes are being introduced:

Bento full record links are occasionally used as permalinks, but we will
be turning them off once we launch Primo.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/IMP-2007

How this addresses that need:

This adds logic to the record controller, toggled by a feature flag,
that makes an Alma SRU query for Aleph record and determines whether we
can unambiguously link to a corresponding record in Primo (i.e., the SRU
returns one and only one record). If we can, we redirect users to that
Primo record, and if not, we redirect them to a generic splash page that
describes the switch from EDS to Primo.

Side effects of this change:

We now toggle feature flags in the test environment using Flipflop's
built-in :test strategy.