Handle irregular Alma openurl links
Why these changes are being introduced:

The app currently expects all Alma openurls to use the 'na06'
Ex Libris server. This could cause problems if we receive records that
use a different server for their openurls.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/IMP-2047

How this addresses that need:

* Logs all records that use a server other than na06 for their Alma
openurl.
* Returns the regular Alma openurl in the record instead of constructing
a Primo openurl for records that don't use the na06 server.

Side effects of this change:

* The expected Alma base URL has been abstracted to ENV, in case we need
to quickly change it in the future.
* Testing this logic required an additional customization to the
missing_fields_primo_articles cassette. This is already a customized
cassette, and the additional customization has been noted in the
normalize_primo_articles model test.