Fix 'view more results' URLs when Primo searches are enabled
Why these changes are being introduced:

The 'view more' links for website/Google and ArchivesSpace searches break
when the Primo feature flag is toggled due to a flawed conditional in
the view_all_links partial.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/IMP-2240

How this addresses that need:

This adjusts the aforementioned conditional to check if the Primo feature
flag is enabled AND if the target param is equal to one of the two Primo
targets.

Side effects of this change:

None.