Display dedup status in result view
Why these changes are being introduced:

Since launch, we've enabled FRBR/dedup records in Primo. We would
also like to display this information in Bento.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/IMP-2262

How this addresses that need:

This checks the FRBR/dedup status of catalog records. If a record
is dedup, the result view renders a stub record with a link to the other
versions instead of RTA.

Side effects of this change:

While debugging a failing integration test, we identified some
problematic logic in the search controller: the controller validates
targets, which makes it impossible to fake target values in tests,
and it checks whether the target in params matches the ENV target value.

We removed the logic matching params to ENV, but as a workaround for
the validator, we added the faked targets we use in tests to the 'valid'
targets array.

A ticket to clean up this logic has been opened:
https://mitlibraries.atlassian.net/browse/IMP-2283