Remove target validation from Search Controller
Why these changes are being introduced:

The Search Controller contains some confusing logic with regard to
validating the target param. The controller checks the target param
against an array of 'valid' targets; if the target is not
considered valid, the controller returns a redirect to the root URL.

This is problematic for a few reasons:

1. It's not very maintainable. We will have to modify the hardcoded
array of valid targets whenever the value of the param, which is stored
in ENV, changes. We've already encountered this with faked param values
in our test suite.
2. It's not very usable. If a user enters a bad param, redirecting
them to the root URL doesn't tell them anything useful about what
went wrong.
3. It's not very intuitive. We already have a method that checks for
which type of search to run based on the target provided. This seems
like the logical place to check for bad targets, rather than three
separate places as we do now.
4. It's not very necessary. This is what 404 errors are for!

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/IMP-2283

How this addresses that need:

This removes the aforementioned target validation logic from the
Search Controller, and instead returns a 404 error (and renders our
custom 404 template) when it gets a bad target param. Note that this
only occurs from the search_boxed view. The bento view continues to
ignore additional params, target or otherwise.

Side effects of this change:

A custom error is used here to avoid a DoubleRenderError.