Remove EDS and other legacy code from Bento
Why these changes are being introduced:

With the launch of Alma and Primo, we no longer need the EDS code
in Bento, and removing it now with save us a major headache in the
future. It also make sense at this point to remove other features
we no longer use, like the button makers, Aleph RTA, etc.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/IMP-2284

How this addresses that need:

This removes several features, including:

* All EDS logic (normalization and search models, as well as code
from the search and record controllers and views, and result model
and views);
* Full records, with the exception of some logic required for Primo
redirects;
* Aleph real-time availability;
* SFX handler;
* Button makers;
* Rainbows;
* Pagination; and
* Authentication.

Side effects of this change:

* Unused VCR cassettes have been removed.
* The system diagram has been updated.
* Some irrelevant assets have been removed (Aleph RTA JS and SCSS
for removed features).
* SearchController#search no longer has test coverage. This method was
actually never tested, but coveralls marked it as such because it
falsely identified a pagination test as coverage. In order to test
test the method, we'd need system tests, which are beyond the scope
of this ticket and may be out of scope of this app, since minimal
click testing can achieve the same result.
* TrackLinks JS, a remnant from Google Analytics, has been removed.
* All env has been removed from app.json, per an EngX team decision.