Guard against nil :an param in record controller
Why these changes are being introduced:

We currently only guard against :an params that don't start with 'mit',
which means the application throws a NoMethodError if that param is nil.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ENGX-102

How this addresses that need:

This adds a guard clause to RecordController#alma_sru to check if
params[:an] is present.

Side effects of this change:

None.