Bugfix for 'Details and requests' links
Why these changes are being introduced:

The November release of Primo VE unexpectedly broke our 'Details
and requests' links, which are constructed as Primo fulldisplay
'deep links'. In the latest release, Ex Libris made two params
(`context` and `search_scope`) required for fulldisplay links.

This was most likely accidental as the change is not mentioned in
the release notes, and the params are not required consistently:
i.e., `context` is required only for local records, and
`search_scope` is required only for CDI records.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/IMP-2358

How this addresses that need:

This adds to our 'Details and requests' links all of the params
included in the Ex Libris docs' fulldisplay example link:
https://developers.exlibrisgroup.com/primo/apis/deep-links-new-ui/

In reality, not all of these params are required, but hopefully
including them now will prevent more surprises later.

Side effects of this change:

I refactored the link method to use `to_query`, which is a bit more
readable to me.