Bugfix for Primo openurls
Why these changes are being introduced:

Some Primo openurls do not resolve correctly. The Alma/Primo team
suspects that the cause is the ctx params in the URL.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ES-572

How this addresses that need:

This removes the ctx params from the Primo openurls.

Side effects of this change:

None.