Fix source param for TIMDEX V2 searches
Why these changes are being introduced:

In V2, all facets have been renamed so it is clearer that they are
facets. This means that V2 searches in Bento will fail, because
the search string uses `source` instead of `sourceFacet`.

More information in this commit:
https://github.com/MITLibraries/timdex/commit/3f2801b6644eca83ac84ae678864f07b4f9c75e9

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/TIMX-170

How this addresses that need:

This adds a timdex_v2 feature to Flipflop, which is used to alter the
search string to correspond with the target TIMDEX version. When this
feature is enabled, `source` becomes `sourceFacet`.

Side effects of this change:

* In order to make both V1 and V2 searches succeed, I had to run
each set of tests separately as we do in TIMDEX. Unfortunately,
this means reverting from the shared Ruby CI to a custom CI, at
least until we shut off V1.
* ClimateControl has been added to the test group to accommodate
testing.