Update Bento to use TIMDEX V2 fields
Why these changes are being introduced:

Bento currently uses deprecated fields in its TIMDEX searches:

* `identifier`
* `publicationDate`
* `source`

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/TIMX-174

How this addresses that need:

This makes the following V1 to V2 replacements:

* `identifier` with `identifiers` (`kind` == 'Collection Identifer')
* `publicationDate` with `dates` (there is some nuance to this;
see the changes to the NormalizeTimdex model for details)
* `source` with `sourceFacet`

Side effects of this change:

* Some cassettes have been regenerated to use the new search params.
* Tests that anticipate results from non-ASpace sources have been
removed. These tests would fail if we were to regenerate cassettes
in the current codebase, since the ASpace source is hardcoded in the
search string.
* Some test assertions have been modified to match the latest
TIMDEX data load.