Better error handling for dates
Why these changes are being introduced:

We encountered a bug after V2 launch in the normalization process,
during which Bento hits an unhandled nil value that breaks the
search results. This is because we are assuming every ASpace
record has a creation date, when in fact some of them have
publication dates instead.

Somewhat related to this is a transmogrifier bug that causes a
`kind` to be associated with a date even when there is no date.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/TIMX-211
* https://mitlibraries.atlassian.net/browse/TIMX-212

How this addresses that need:

This updates `NormalizeTimdex#extract_dates` to select as relevant
dates of either the 'creation' _or_ 'publication' date kind. It
also adds a guard clause to return if the `relevant_dates` array
is empty (a precaution that would have prevented this error).

Side effects of this change:

I removed a superfluous check for `date['kind'] == 'creation'` in
the guard clause in the `extract_dates` method that checks for
a date range. At this point in the method, we have already
confirmed that the relevant date is either a creation or publication
date.