Handle nil dates in TIMDEX searches
Why these changes are being introduced:

It's possible that a result may have a `dates` value
of `nil`. This is not currently handled in the
TIMDEX normalization model.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/TIMX-219

How this addresses that need:

This adds a guard clause for nil dates to NormalizeTimdex#extract_dates.

Side effects of this change:

Some style changes were made by rubocop to unrelated tests.