Changes from pairing session
Stephanie and I paired today to fine-tune the design a bit. The
goals of the session were to fix the padding, reduce the font size
in large viewports, and ensure that the text didn't wrap beyond
two lines in smaller viewports.

We made the following changes:

* Removes `copy-lead` and `box-content` classes.
* Sets padding to `20px 10px`.
* Sets font-size to `$fs-large` (20px) at viewport widths larger
than our medium breakpoint, and `$fs-base` (16px) at smaller widths.
* Explicitly sets `display: block`. (This rule is implied/inherited,
but it felt useful to declare it because inline-block does weird
things.)

Other things we tried but decided not to do:

* Put `display: block` and `text-wrap: nowrap` on the link in
smaller viewports. This mostly looked good, but squeezed the
spacing from the sides.
* Leave the font size as `$fs-large` in smaller viewports. We were
able to do this without wrapping by reducing the padding, but it
was visually out of proportion.

Co-authored-by: Stephanie Hartman <hartman@mit.edu>