Replace Hints with Suggested Resources
Why these changes are being introduced:

We've replaced the Hints system built into Bento with the
Suggested Resources detector, provided by TACOS API. Since Bento is
already logging search terms to TACOS in production, it makes
sense to switch to using Suggested Resources.

Relevant ticket(s):

* [TCO-66](https://mitlibraries.atlassian.net/browse/TCO-66)

How this addresses that need:

This removes the Hint feature and replaces it with Suggested
Resources. We are currently triggering Suggested Resources in
the same way as Hints (and the other controllers), but it might
make sense to refactor this feature using a tool like Stimulus or
Turbo. I've opened (ENGX-283)[https://mitlibraries.atlassian.net/browse/ENGX-283]
to track further work on this.

While discussing this feature, we realized that GraphQL returns
an array for `suggestedResources` even though that detector will
only return a single resource. We are handling this in Bento by
just calling `first` on the array, but we should reconsider whether
the GraphQL type is returning the correct data structure. The spike
ticket [TCO-66](https://mitlibraries.atlassian.net/browse/TCO-66)
will further explore this problem space.

Side effects of this change:

Debug and ClimateControl have been added to the Gemfile. The
former is what we should be using to debug instead of Pry, and
the latter was needed to test various ENV conditions.

Co-authored-by: Jeremy Prevost <JPrevost@users.noreply.github.com>