Remove sqlite version requirement
I had temporarily prevented sqlite from installing because it will fail until we upgrade Rails. Because we've enabled the frozen config in bundler, Heroku builds won't work until this is either removed from Gemfile.lock or added to the Gemfile. I'd prefer not to add it to the Gemfile, because we might then forget to upgrade the gem when we upgrade Rails.